Moved the dart files into screens folder
