+++ kaleidoscope_collaborative/lib/screens/first_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'constants.dart';
-import "login_screen.dart";
+import 'login_screen.dart';
 import 'signup_screen.dart';
 
 class FirstScreen extends StatefulWidget{
