+++ kaleidoscope_collaborative/lib/screens/login_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:kaleidoscope_collaborative/constants.dart';
+import 'package:kaleidoscope_collaborative/screens/constants.dart';
 
 
 class LoginScreen extends StatefulWidget{
